requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-openai
langchain-community
langchain-huggingface
langchain-groq
langchain-text-splitters
sentence-transformers
faiss-cpu
PyMuPDF
openai
groq
python-dotenv
pydantic-settings
requests
tqdm